Split the file
  ref #9
